--- cmd/graceful_shutdown.go
@@ -1,62 +1,0 @@
-package main
-
-import (
-	"context"
-	"os"
-	"os/signal"
-	"sync"
-	"syscall"
-	"time"
-
-	"go.uber.org/zap"
-)
-
-type operation func(ctx context.Context) error
-
-func gracefulShutdown(ctx context.Context, timeout time.Duration, log *zap.Logger, ops map[string]operation) <-chan struct{} {
-	wait := make(chan struct{})
-	go func() {
-		s := make(chan os.Signal, 1)
-
-		signal.Notify(s, syscall.SIGINT, syscall.SIGTERM, os.Interrupt)
-		sig := <-s
-
-		log.Named("graceful shutdown").Sugar().
-			Infof("got signal \"%v\" shutting down service", sig)
-
-		timeoutFunc := time.AfterFunc(timeout, func() {
-			log.Named("graceful shutdown").Sugar().
-				Errorf("timeout %v ms has been elapsed, force exit", timeout.Milliseconds())
-			os.Exit(0)
-		})
-
-		defer timeoutFunc.Stop()
-
-		var wg sync.WaitGroup
-
-		for key, op := range ops {
-			wg.Add(1)
-			innerOp := op
-			innerKey := key
-			go func() {
-				defer wg.Done()
-
-				log.Named("graceful shutdown").Sugar().
-					Infof("cleaning up: %v", innerKey)
-				if err := innerOp(ctx); err != nil {
-					log.Named("graceful shutdown").Sugar().
-						Errorf("%v: clean up failed: %v", innerKey, err.Error())
-					return
-				}
-
-				log.Named("graceful shutdown").Sugar().
-					Infof("%v was shutdown gracefully", innerKey)
-			}()
-		}
-
-		wg.Wait()
-		close(wait)
-	}()
-
-	return wait
-}
